+++ src/main/java/dev/hoyin1600p/vhaccelerator/mixin/backport/modernfix/recipe/TagValueAccessor.java
@@ -1,0 +1,25 @@
+/*
+ * SPDX-License-Identifier: LGPL-3.0-or-later
+ *
+ * Adapted for VH Accelerator from ModernFix.
+ * Upstream repository: https://github.com/embeddedt/ModernFix
+ * Upstream source: forge/src/main/java/org/embeddedt/modernfix/forge/mixin/perf/faster_ingredients/IngredientMixin.java
+ * Upstream commit: 5048e74c79724029a8ba2e3d7cdd37c83320ad4e
+ * Original copyright: Copyright (c) 2025 embeddedt and ModernFix contributors
+ * VH Accelerator modifications: Copyright (C) 2026 HoYin1600p
+ * Modified: 2026-08-30; replaces a newer access widener with a 1.18.2 Mixin
+ * accessor for the vanilla tag key.
+ */
+package dev.hoyin1600p.vhaccelerator.mixin.backport.modernfix.recipe;
+
+import net.minecraft.tags.TagKey;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.crafting.Ingredient;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+@Mixin(Ingredient.TagValue.class)
+public interface TagValueAccessor {
+    @Accessor("tag")
+    TagKey<Item> vha$getTag();
+}